+++ CHANGELOG.md
@@ -4,14 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [1.1.9 - 2023-10-17]
+## [1.1.9 - 2023-10-20]
 ### Fixed
 - Finalize report generation correctly
 - Read files with configuration charset
 - Write files with configuration charset
 
 ### Added
-- Expiration date in hours
+- Export expiration in days
 
 ### Changed
 - FHIR_PATH to FHIR_SEARCH